+++ be-yelpcamp/package.json
@@ -5,6 +5,7 @@
     "cors": "^2.8.5",
     "cros": "^1.0.1",
     "express": "^4.21.2",
+    "express-session": "^1.18.1",
     "joi": "^17.13.3",
     "mongoose": "^8.8.4"
   },
